Deprecate sunpy.io.fits
