+++ setup.cfg
@@ -129,6 +129,8 @@ filterwarnings =
     error
     # Do not fail on pytest config issues (i.e. missing plugins) but do show them
     always::pytest.PytestConfigWarning
+    # Catch deprecation warning when collecting sunpy.io.fits
+    ignore:The `sunpy.io.fits` module is deprecated::sunpy.io.fits
     #
     # A list of warnings to ignore follows. If you add to this list, you MUST
     # add a comment or ideally a link to an issue that explains why the warning
